+++ project_app/components/ButtonGroup.js
@@ -1,0 +1,22 @@
+import React from "react";
+
+function ButtonGroup() {
+  return (
+    <div>
+      <IconButton label="Events">
+        <p>Events</p>
+      </IconButton>
+      <IconButton label="FAQ">
+        <p>FAQ</p>
+      </IconButton>
+      <IconButton label="About Us">
+        <p>About Us</p>
+      </IconButton>
+      <IconButton label="Game night">
+        <p>Game Night</p>
+      </IconButton>
+    </div>
+  );
+}
+
+export default ButtonGroup;
