+++ project_app/package.json
@@ -23,6 +23,7 @@
   },
   "devDependencies": {
     "@svgr/webpack": "^6.5.1",
+    "@types/styled-components": "^5.1.26",
     "eslint": "8.32.0",
     "eslint-config-next": "13.1.2",
     "jest": "^29.3.1"
