--- project_app/pages/HomePage.js
@@ -1,12 +1,0 @@
-import React from "react";
-import Layout from "../components/Layout";
-
-function HomePage() {
-  return (
-    <Layout>
-      <div>HomePage</div>
-    </Layout>
-  );
-}
-
-export default HomePage;
