+++ project_app/pages/index.tsx
@@ -1,0 +1,8 @@
+import Layout from "../components/Layout";
+
+export default function Home() {
+  return (
+    <Layout/>
+    
+  );
+}
